Create Concatenated Sum
it is  method named checkContenatedSum that tests if a 
number has this generalized property. The signature of the method is int 
checkConcatenatedSum(int n, int catlen) where n is the number and catlen is the number of 
times to concatenate each digit before summing. 
The method returns 1 if n is equal to the sum of each of its digits contenated catlen times. 
Otherwise, it returns 0